fix: rename testutils and add more descriptive errors in testing
